#18 Rename and rearrange models, Fix tests
