+++ build.gradle.kts
@@ -2,13 +2,13 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 import java.net.URI
 
 val kotlinVersion = "1.3.21"
-val http4kVersion = "3.113.0"
+val http4kVersion = "3.132.0"
 val log4jVersion = "2.11.2"
 val jacksonVersion = "2.9.8"
 
 plugins {
     kotlin("jvm") version "1.3.21"
-    id("com.github.ben-manes.versions") version "0.20.0"
+    id("com.github.ben-manes.versions") version "0.21.0"
 }
 
 repositories {
@@ -23,7 +23,7 @@ dependencies {
     implementation("org.http4k:http4k-server-jetty:$http4kVersion")
     implementation("org.http4k:http4k-format-jackson:$http4kVersion")
     implementation("org.http4k:http4k-client-apache:$http4kVersion")
-    implementation("org.jetbrains.exposed:exposed:0.12.2")
+    implementation("org.jetbrains.exposed:exposed:0.13.6")
     implementation("io.jsonwebtoken:jjwt:0.9.1")
     implementation("com.h2database:h2:1.4.198")
     implementation("org.apache.logging.log4j:log4j-core:$log4jVersion")
@@ -33,8 +33,8 @@ dependencies {
     implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:$jacksonVersion")
     implementation("com.fasterxml.jackson.datatype:jackson-datatype-joda:$jacksonVersion")
 
-    testImplementation("io.kotlintest:kotlintest-runner-junit5:3.3.0-RC3")
-    testImplementation("io.mockk:mockk:1.9.1")
+    testImplementation("io.kotlintest:kotlintest-runner-junit5:3.3.2")
+    testImplementation("io.mockk:mockk:1.9.3")
 }
 
 tasks.test {
