+++ src/main/kotlin/conduit/handler/GetArticleCommentsHandler.kt
@@ -2,11 +2,11 @@ package conduit.handler
 
 import conduit.model.ArticleSlug
 import conduit.model.CommentDto
-import conduit.model.extractEmail
 import conduit.model.toProfile
 import conduit.repository.ConduitTxManager
 import conduit.util.HttpException
 import conduit.util.TokenAuth
+import conduit.util.extractEmail
 import org.http4k.core.Status
 
 interface GetArticleCommentsHandler {
