+++ src/main/kotlin/conduit/handler/GetArticlesFeedHandler.kt
@@ -1,11 +1,11 @@
 package conduit.handler
 
 import conduit.model.ArticleDto
-import conduit.model.extractEmail
 import conduit.model.toProfile
 import conduit.repository.ConduitTxManager
 import conduit.util.HttpException
 import conduit.util.TokenAuth
+import conduit.util.extractEmail
 import org.http4k.core.Status
 
 interface GetArticlesFeedHandler {
