+++ src/main/kotlin/conduit/handler/GetCurrentUserHandler.kt
@@ -1,9 +1,9 @@
 package conduit.handler
 
-import conduit.model.extractEmail
 import conduit.repository.ConduitTxManager
 import conduit.util.HttpException
 import conduit.util.TokenAuth
+import conduit.util.extractEmail
 import org.http4k.core.Status
 
 interface GetCurrentUserHandler {
