+++ src/main/kotlin/conduit/handler/GetProfileHandler.kt
@@ -2,10 +2,10 @@ package conduit.handler
 
 import conduit.model.Profile
 import conduit.model.Username
-import conduit.model.extractEmail
 import conduit.repository.ConduitTxManager
 import conduit.util.HttpException
 import conduit.util.TokenAuth
+import conduit.util.extractEmail
 import org.http4k.core.Status
 
 interface GetProfileHandler {
