+++ src/main/kotlin/conduit/handler/UpdateArticleHandler.kt
@@ -1,9 +1,13 @@
 package conduit.handler
 
-import conduit.model.*
+import conduit.model.ArticleDto
+import conduit.model.ArticleSlug
+import conduit.model.UpdateArticle
+import conduit.model.toProfile
 import conduit.repository.ConduitTxManager
 import conduit.util.HttpException
 import conduit.util.TokenAuth
+import conduit.util.extractEmail
 import org.http4k.core.Status
 
 interface UpdateArticleHandler {
